+++ pgdrive/__init__.py
@@ -1,2 +1,3 @@
 import pgdrive.register
-from pgdrive.envs import PGDriveEnv, GeneralizationRacing, ActionRepeat, ChangeFrictionEnv, RemotePGDrive
+from pgdrive.envs import PGDriveEnv, GeneralizationRacing, ActionRepeat, ChangeFrictionEnv, RemotePGDrive, \
+    TopDownPGDriveEnv, TopDownSingleFramePGDriveEnv
