+++ pgdrive/envs/__init__.py
@@ -3,6 +3,7 @@ from pgdrive.envs.generation_envs.change_density_env import ChangeDensityEnv
 from pgdrive.envs.generation_envs.change_friction_env import ChangeFrictionEnv
 from pgdrive.envs.pgdrive_env import PGDriveEnv
 from pgdrive.envs.remote_pgdrive_env import RemotePGDrive
+from pgdrive.envs.top_down_env import TopDownSingleFramePGDriveEnv, TopDownPGDriveEnv
 
 # For compatibility
 GeneralizationRacing = PGDriveEnv
