+++ pgdrive/world/top_down_observation/__init__.py
@@ -2,3 +2,5 @@
 We implement this top-down view render based on eleurent/Highway-Env
 See more information on its Github page: https://github.com/eleurent/highway-env
 """
+from pgdrive.world.top_down_observation.top_down_multi_channel import TopDownMultiChannel
+from pgdrive.world.top_down_observation.top_down_observation import TopDownObservation
