+++ pgdrive/world/top_down_observation/top_down_multi_channel.py
@@ -1,6 +1,7 @@
 import math
 from collections import deque
 
+import cv2
 import gym
 import numpy as np
 from pgdrive.scene_creator.ego_vehicle.base_vehicle import BaseVehicle
@@ -9,7 +10,6 @@ from pgdrive.utils.constans import Decoration
 from pgdrive.world.top_down_observation.top_down_obs_impl import WorldSurface, COLOR_BLACK, \
     VehicleGraphics, LaneGraphics, ObservationWindowMultiChannel
 from pgdrive.world.top_down_observation.top_down_observation import TopDownObservation
-import cv2
 
 pygame = import_pygame()
 COLOR_WHITE = pygame.Color("white")
@@ -27,7 +27,7 @@ class TopDownMultiChannel(TopDownObservation):
     CHANNEL_NAMES = ["road_network", "traffic_flow", "target_vehicle", "navigation", "past_pos"]
 
     def __init__(
-        self, vehicle_config, env, clip_rgb: bool, frame_stack: int = 5, post_stack: int = 5, frame_skip: int = 5
+            self, vehicle_config, env, clip_rgb: bool, frame_stack: int = 5, post_stack: int = 5, frame_skip: int = 5
     ):
         super(TopDownMultiChannel, self).__init__(vehicle_config, env, clip_rgb)
         self.num_stacks = 9
@@ -220,7 +220,7 @@ class TopDownMultiChannel(TopDownObservation):
 
     @property
     def observation_space(self):
-        shape = self.obs_shape + (self.num_stacks, )
+        shape = self.obs_shape + (self.num_stacks,)
         if self.rgb_clip:
             return gym.spaces.Box(-0.0, 1.0, shape=shape, dtype=np.float32)
         else:
